+++ upload/modules/rees46/controllers/front/cron.php
@@ -100,7 +100,7 @@ class Rees46CronModuleFrontController extends ModuleFrontController
                 $xml .= "\n" . '      <currency id="' . $currency['iso_code'] . '" rate="1"/>';
             } elseif ($currency['active'] == 1) {
                 $xml .= "\n" . '      <currency id="' . $currency['iso_code'] . '" ';
-                $xml .= 'rate="' . number_format(1 / $currency['conversion_rate'], 4, '.', '') . '"/>';
+                $xml .= 'rate="' . number_format(1 / $currencies[]['conversion_rate'], 4, '.', '') . '"/>';
             }
         }
 
@@ -166,7 +166,12 @@ class Rees46CronModuleFrontController extends ModuleFrontController
                 $xml .= ' available="true">' . "\n";
                 $xml .= '        <url>'.$this->replacer($this->context->link->getProductLink($product->id)).'</url>' . "\n";
 
-                $price = $product->getPrice(!Tax::excludeTaxeOption());
+                if (Configuration::get('REES46_XML_PRICE')) {
+                    $price = $product->getPrice(!Tax::excludeTaxeOption());
+                } else {
+                    $price = $product->getPrice(Tax::excludeTaxeOption());
+                }
+
                 $currency = new Currency((int)Configuration::get('REES46_XML_CURRENCY'));
                 $price *= $currency->conversion_rate;